+++ test/examples/flappy_bird/browser-entry/browser-entry.visualization.utils.ts
@@ -1,7 +1,5 @@
 import Network from '../../../../src/architecture/network';
 import {
-  FLAPPY_CENTER_BLUE_RAMP,
-  FLAPPY_LIGHT_NEON_RAMP,
   FLAPPY_MONOSPACE_FONT_FAMILY,
   FLAPPY_NEON_PALETTE,
   FLAPPY_NETWORK_HEADER_FONT_SIZE_PX,
@@ -37,13 +35,11 @@ import {
   FLAPPY_NETWORK_NODE_LABEL_SIZE_RATIO,
   FLAPPY_NETWORK_OUTPUT_NODE_GLOW_COLOR,
   FLAPPY_NETWORK_OUTPUT_NODE_STROKE_COLOR,
-  FLAPPY_REGULAR_NEON_RAMP,
   FLAPPY_VIEWPORT_NETWORK_OVERLAY_HIDDEN_BREAKPOINT_PX,
 } from '../constants/constants';
 import { applyAlphaToHexColor, clamp } from './browser-entry.math.utils';
 import type {
   ColorLegendRow,
-  ColorTier,
   NetworkLegendLayout,
   NetworkNodeDimensionsLike,
   PositionedNetworkNodeLike,
@@ -60,93 +56,20 @@ import {
   FLAPPY_NETWORK_DOTTED_CONNECTION_STEP_COMPACT_RATIO,
   FLAPPY_NETWORK_DOTTED_CONNECTION_WIDTH_SPACING_RATIO,
 } from './visualization/visualization.constants';
+import { assertFiniteLegendBound } from './visualization/visualization.errors';
+import {
+  resolveNetworkVisualizationColorScales,
+  resolveTierColor,
+} from './visualization/visualization.colors.utils';
+import { formatNodeBiasLabel } from './visualization/visualization.topology.utils';
 
-/**
- * Builds logarithmic diverging color tiers with a center band and edge extension.
- *
- * @param input - Tier creation options.
- * @returns Ordered tier list.
- */
-export function createLogDivergingColorTiers(input: {
-  maxAbsValue: number;
-  centerBlueThreshold: number;
-  negativePalette: readonly string[];
-  centerBluePalette: readonly string[];
-  positivePalette: readonly string[];
-  logarithmicSteepness: number;
-  edgeStartAbsValue?: number;
-  edgeTierCount?: number;
-}): ColorTier[] {
-  // Step 1: Resolve mirrored magnitude tracks for negative and positive ranges.
-  const negativeMagnitudes = resolveTwoZoneMagnitudes({
-    minimumMagnitude: input.centerBlueThreshold,
-    maximumMagnitude: input.maxAbsValue,
-    tierCount: input.negativePalette.length,
-    logarithmicSteepness: input.logarithmicSteepness,
-    edgeStartAbsValue: input.edgeStartAbsValue,
-    edgeTierCount: input.edgeTierCount,
-  });
-
-  const positiveMagnitudes = resolveTwoZoneMagnitudes({
-    minimumMagnitude: input.centerBlueThreshold,
-    maximumMagnitude: input.maxAbsValue,
-    tierCount: input.positivePalette.length,
-    logarithmicSteepness: input.logarithmicSteepness,
-    edgeStartAbsValue: input.edgeStartAbsValue,
-    edgeTierCount: input.edgeTierCount,
-  });
-
-  // Step 2: Build negative, center, and positive tier groups.
-  const negativeTiers = input.negativePalette.map((color, colorIndex) => {
-    const magnitude = negativeMagnitudes[colorIndex];
-
-    return {
-      upperBound: -magnitude,
-      color,
-    };
-  });
-
-  const centerTiers = input.centerBluePalette.map((color, colorIndex) => {
-    const linearProgress = (colorIndex + 1) / input.centerBluePalette.length;
-    return {
-      upperBound:
-        -input.centerBlueThreshold +
-        linearProgress * input.centerBlueThreshold * 2,
-      color,
-    };
-  });
-
-  const positiveTiers = input.positivePalette.map((color, colorIndex) => {
-    const magnitude = positiveMagnitudes[colorIndex];
-
-    return {
-      upperBound: magnitude,
-      color,
-    };
-  });
-
-  // Step 3: Merge and sort tiers by ascending upper bounds.
-  return [...negativeTiers, ...centerTiers, ...positiveTiers].toSorted(
-    (leftTier, rightTier) => leftTier.upperBound - rightTier.upperBound,
-  );
-}
-
-/**
- * Resolves a color from ordered tier definitions.
- *
- * @param value - Numeric value to classify.
- * @param tiers - Ordered tier list.
- * @param aboveTierColor - Fallback color for values above the last tier.
- * @returns Resolved color string.
- */
-export function resolveTierColor(
-  value: number,
-  tiers: ColorTier[],
-  aboveTierColor: string,
-): string {
-  const resolvedTier = tiers.find((tier) => value <= tier.upperBound);
-  return resolvedTier?.color ?? aboveTierColor;
-}
+export {
+  createLogDivergingColorTiers,
+  resolveBiasRangeColor,
+  resolveConnectionRangeColor,
+  resolveNetworkVisualizationColorScales,
+  resolveTierColor,
+} from './visualization/visualization.colors.utils';
 
 /**
  * Creates legend rows from ordered tiers.
@@ -255,7 +178,7 @@ export function resolveNetworkLegendLayout(
  * @param positionByNodeIndex - Node layout map.
  * @returns Nothing.
  */
-export function drawWeightedConnectionsLayer(
+export function drawWeightedConnectionsLayerInternal(
   context: CanvasRenderingContext2D,
   runtimeConnections: VisualNetworkConnectionLike[],
   positionByNodeIndex: Map<number, PositionedNetworkNodeLike>,
@@ -401,7 +324,7 @@ function drawSquareDottedConnection(
  * @param nodeDimensions - Node dimensions.
  * @returns Nothing.
  */
-export function drawBiasNodesLayer(
+export function drawBiasNodesLayerInternal(
   context: CanvasRenderingContext2D,
   positionedNodes: PositionedNetworkNodeLike[],
   nodeDimensions: NetworkNodeDimensionsLike,
@@ -504,7 +427,7 @@ export function drawBiasNodesLayer(
  * @param architectureLabel - Header label.
  * @returns Nothing.
  */
-export function drawNetworkVisualizationHeader(
+export function drawNetworkVisualizationHeaderInternal(
   context: CanvasRenderingContext2D,
   architectureLabel: string,
 ): void {
@@ -524,7 +447,7 @@ export function drawNetworkVisualizationHeader(
  * @param context - Render context.
  * @returns Nothing.
  */
-export function drawNetworkColorLegend(
+export function drawNetworkColorLegendInternal(
   context: CanvasRenderingContext2D,
   architectureLabel: string,
   colorScales: NetworkVisualizationColorScales,
@@ -667,7 +590,7 @@ export function drawNetworkColorLegend(
  * @param context - Render context.
  * @returns Legend layout.
  */
-export function resolveDefaultNetworkLegendLayout(
+export function resolveDefaultNetworkLegendLayoutInternal(
   context: CanvasRenderingContext2D,
   network: Network | undefined,
 ): NetworkLegendLayout {
@@ -688,193 +611,12 @@ export function resolveDefaultNetworkLegendLayout(
 }
 
 /**
- * Resolves connection color for a raw weight.
- *
- * @param connectionWeight - Connection weight.
- * @returns Tier color.
- */
-export function resolveConnectionRangeColor(connectionWeight: number): string {
-  const connectionScale = createDynamicColorScale([connectionWeight], {
-    minimumValue: -1,
-    maximumValue: 1,
-  });
-  return resolveTierColor(
-    connectionWeight,
-    connectionScale.tiers,
-    connectionScale.aboveTierColor,
-  );
-}
-
-/**
- * Resolves bias color for a raw node bias.
- *
- * @param nodeBias - Node bias.
- * @returns Tier color.
- */
-export function resolveBiasRangeColor(nodeBias: number): string {
-  const biasScale = createDynamicColorScale([nodeBias], {
-    minimumValue: -1,
-    maximumValue: 1,
-  });
-  return resolveTierColor(nodeBias, biasScale.tiers, biasScale.aboveTierColor);
-}
-
-/**
- * Resolves dynamic connection/bias color scales from the active network range.
- *
- * @param network - Active network.
- * @returns Dynamic scales used by graph drawing and legend rows.
- */
-export function resolveNetworkVisualizationColorScales(
-  network: Network | undefined,
-): NetworkVisualizationColorScales {
-  const connectionValues = ((network?.connections ?? []) as Array<{
-    weight?: number;
-  }>)
-    .map((connection) => Number(connection.weight ?? 0))
-    .filter((weight) => Number.isFinite(weight));
-
-  const biasValues = ((network?.nodes ?? []) as Array<{
-    type?: string;
-    bias?: number;
-  }>)
-    .filter((node) => node.type !== 'output')
-    .map((node) => Number(node.bias ?? 0))
-    .filter((bias) => Number.isFinite(bias));
-
-  return {
-    connectionScale: createDynamicColorScale(connectionValues, {
-      minimumValue: -1,
-      maximumValue: 1,
-    }),
-    biasScale: createDynamicColorScale(biasValues, {
-      minimumValue: -1,
-      maximumValue: 1,
-    }),
-  };
-}
-
-/**
- * Creates a linear color scale that spans the observed value range.
- *
- * @param values - Runtime values to map.
- * @param fallbackRange - Fallback range when values are empty/non-finite.
- * @returns Dynamic color scale.
- */
-function createDynamicColorScale(
-  values: number[],
-  fallbackRange: { minimumValue: number; maximumValue: number },
-): DynamicColorScale {
-  const finiteValues = values.filter((value) => Number.isFinite(value));
-  const observedMinimumValue =
-    finiteValues.length > 0
-      ? Math.min(...finiteValues)
-      : fallbackRange.minimumValue;
-  const observedMaximumValue =
-    finiteValues.length > 0
-      ? Math.max(...finiteValues)
-      : fallbackRange.maximumValue;
-  const hasRange = observedMaximumValue > observedMinimumValue;
-
-  const minimumValue = hasRange
-    ? observedMinimumValue
-    : observedMinimumValue - Math.max(1e-6, Math.abs(observedMinimumValue) * 0.01);
-  const maximumValue = hasRange
-    ? observedMaximumValue
-    : observedMaximumValue + Math.max(1e-6, Math.abs(observedMaximumValue) * 0.01);
-
-  const dynamicTiers = resolveSignedDynamicTiers(minimumValue, maximumValue);
-  const aboveTierColor =
-    dynamicTiers.at(-1)?.color ?? FLAPPY_NEON_PALETTE.currentRunText;
-
-  return {
-    minimumValue,
-    maximumValue,
-    tiers: dynamicTiers,
-    aboveTierColor,
-  };
-}
-
-/**
- * Resolves tier colors by sign so negative and positive sides use distinct ramps.
- *
- * @param minimumValue - Observed minimum value.
- * @param maximumValue - Observed maximum value.
- * @returns Ordered dynamic tiers.
- */
-function resolveSignedDynamicTiers(
-  minimumValue: number,
-  maximumValue: number,
-): ColorTier[] {
-  const negativePaletteLowToHigh = [
-    ...FLAPPY_LIGHT_NEON_RAMP.toReversed(),
-    ...FLAPPY_CENTER_BLUE_RAMP,
-  ] as const;
-  const positivePaletteLowToHigh = [
-    ...FLAPPY_CENTER_BLUE_RAMP,
-    ...FLAPPY_REGULAR_NEON_RAMP,
-  ] as const;
-
-  if (minimumValue >= 0) {
-    return createLinearColorTiers({
-      minimumValue,
-      maximumValue,
-      palette: positivePaletteLowToHigh,
-    });
-  }
-
-  if (maximumValue <= 0) {
-    return createLinearColorTiers({
-      minimumValue,
-      maximumValue,
-      palette: negativePaletteLowToHigh,
-    });
-  }
-
-  const negativeTiers = createLinearColorTiers({
-    minimumValue,
-    maximumValue: 0,
-    palette: negativePaletteLowToHigh,
-  });
-  const positiveTiers = createLinearColorTiers({
-    minimumValue: 0,
-    maximumValue,
-    palette: positivePaletteLowToHigh,
-  });
-  return [...negativeTiers, ...positiveTiers];
-}
-
-/**
- * Creates linearly spaced color tiers between observed min and max.
- *
- * @param input - Tier creation options.
- * @returns Ordered tier list.
- */
-function createLinearColorTiers(input: {
-  minimumValue: number;
-  maximumValue: number;
-  palette: readonly string[];
-}): ColorTier[] {
-  const paletteSize = Math.max(1, input.palette.length);
-  const range = Math.max(1e-12, input.maximumValue - input.minimumValue);
-  const step = range / paletteSize;
-
-  return input.palette.map((color, colorIndex) => ({
-    upperBound:
-      colorIndex === paletteSize - 1
-        ? input.maximumValue
-        : input.minimumValue + step * (colorIndex + 1),
-    color,
-  }));
-}
-
-/**
  * Formats node bias labels with fixed sign and precision.
  *
  * @param nodeBias - Node bias value.
  * @returns Label text.
  */
-export function formatNodeBiasLabel(nodeBias: number): string {
+export function formatNodeBiasLabelInternal(nodeBias: number): string {
   const roundedBias = Number.isFinite(nodeBias) ? nodeBias : 0;
   return `${roundedBias >= 0 ? '+' : ''}${roundedBias.toFixed(2)}`;
 }
@@ -1157,101 +899,12 @@ function resolveHiddenNodeDepthByTopology(
 }
 
 /**
- * Resolves magnitudes in near and edge zones.
- *
- * @param input - Magnitude options.
- * @returns Ordered magnitudes.
- */
-function resolveTwoZoneMagnitudes(input: {
-  minimumMagnitude: number;
-  maximumMagnitude: number;
-  tierCount: number;
-  logarithmicSteepness: number;
-  edgeStartAbsValue?: number;
-  edgeTierCount?: number;
-}): number[] {
-  // Step 1: Resolve safe tier counts and edge-zone split configuration.
-  const safeTierCount = Math.max(1, input.tierCount);
-  const targetEdgeStart = clamp(
-    input.edgeStartAbsValue ?? input.maximumMagnitude,
-    input.minimumMagnitude,
-    input.maximumMagnitude,
-  );
-  const requestedEdgeTierCount = clamp(
-    input.edgeTierCount ?? 0,
-    0,
-    safeTierCount,
-  );
-  const edgeTierCount =
-    targetEdgeStart >= input.maximumMagnitude ? 0 : requestedEdgeTierCount;
-  const nearTierCount = Math.max(1, safeTierCount - edgeTierCount);
-
-  // Step 2: Build logarithmic near-zone magnitudes.
-  const nearMagnitudes = Array.from(
-    { length: nearTierCount },
-    (_unusedValue, tierIndex) => {
-      const logarithmicProgress = mapLogarithmicProgress(
-        tierIndex + 1,
-        nearTierCount,
-        input.logarithmicSteepness,
-      );
-      return (
-        input.minimumMagnitude +
-        (targetEdgeStart - input.minimumMagnitude) * logarithmicProgress
-      );
-    },
-  );
-
-  // Step 3: Return early when no edge-zone tiers are requested.
-  if (edgeTierCount === 0) {
-    return nearMagnitudes;
-  }
-
-  // Step 4: Build linear edge-zone magnitudes and concatenate.
-  const edgeMagnitudes = Array.from(
-    { length: edgeTierCount },
-    (_unusedValue, edgeIndex) => {
-      const linearProgress = (edgeIndex + 1) / edgeTierCount;
-      return (
-        targetEdgeStart +
-        (input.maximumMagnitude - targetEdgeStart) * linearProgress
-      );
-    },
-  );
-
-  return [...nearMagnitudes, ...edgeMagnitudes];
-}
-
-/**
- * Maps linear position to logarithmic progress.
- *
- * @param position - 1-based position.
- * @param totalPositions - Total position count.
- * @param logarithmicSteepness - Log curve steepness.
- * @returns Logarithmic progress.
- */
-function mapLogarithmicProgress(
-  position: number,
-  totalPositions: number,
-  logarithmicSteepness: number,
-): number {
-  const normalizedPosition = clamp(
-    position / Math.max(1, totalPositions),
-    0,
-    1,
-  );
-  return (
-    Math.log1p(logarithmicSteepness * normalizedPosition) /
-    Math.log1p(logarithmicSteepness)
-  );
-}
-
-/**
  * Formats a legend bound with fixed precision.
  *
  * @param value - Numeric bound.
  * @returns Formatted bound string.
  */
 function formatLegendBound(value: number): string {
+  assertFiniteLegendBound(value);
   return value.toFixed(2);
 }
