+++ test/examples/flappy_bird/browser-entry/network-view/network-view.constants.ts
@@ -1,0 +1,10 @@
+/**
+ * Ordered labels for grouped Flappy network input bands.
+ */
+export const FLAPPY_INPUT_GROUP_LABELS: readonly string[] = [
+  'CURRENT FRAME',
+  'PREVIOUS FRAME',
+  'TWO FRAMES AGO',
+  'ACT',
+  'RATE',
+] as const;
