+++ test/examples/flappy_bird/browser-entry/network-view/network-view.types.ts
@@ -1,0 +1,10 @@
+/**
+ * Input-group label band geometry and style contract.
+ */
+export interface InputGroupLabelBand {
+  label: string;
+  startNodeIndex: number;
+  endNodeIndex: number;
+  backgroundColor: string;
+  orientation: 'vertical' | 'horizontal';
+}
