+++ test/examples/flappy_bird/browser-entry/visualization/visualization.legend.utils.ts
@@ -1,0 +1,17 @@
+import Network from '../../../../../src/architecture/network';
+import type { NetworkLegendLayout } from '../browser-entry.types';
+import { resolveDefaultNetworkLegendLayoutInternal } from '../browser-entry.visualization.utils';
+
+/**
+ * Resolves default legend layout from internal tier definitions.
+ *
+ * @param context - Render context.
+ * @param network - Active network instance.
+ * @returns Legend layout.
+ */
+export function resolveDefaultNetworkLegendLayout(
+  context: CanvasRenderingContext2D,
+  network: Network | undefined,
+): NetworkLegendLayout {
+  return resolveDefaultNetworkLegendLayoutInternal(context, network);
+}
